Add files via upload
A Python program that analyzes a linguistic database dump file. It analyzes the frequency with which various words have occurred in a given year, frequency with which various letters of the alphabet have occurred in a given year, and gathers statistics on the length of the words and compiles it to create box and whiskers diagram.